requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0

# OpenAI API
openai==1.12.0

# HTTP client
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Document generation (for export functionality)
python-docx==1.1.0
reportlab==4.0.7
fpdf2==2.8.3

# Date and time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production server (optional)
gunicorn==21.2.0 